# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Badal API
  version: 0.12.0
  description: |
    API REST versionnee de la plateforme Badal.

    ## Conventions

    - Toutes les reponses sont en JSON et portent l'en-tete `X-Request-ID`, genere par le serveur
      (jamais repris du client). Le meme identifiant figure dans le champ `request_id` des erreurs
      et dans les journaux : le communiquer au support.
    - La langue des messages se choisit avec l'en-tete `Accept-Language` (`fr` ou `ar`, `fr` par
      defaut). Seul le champ `message` est traduit ; `code` et les cles de `errors` sont stables.
    - Les erreurs suivent toutes le schema `Error` (voir `components/responses`). Les clients
      s'appuient sur `code`, jamais sur `message`.
    - Une limite de debit est appliquee par route ; en cas de depassement la reponse est `429` avec
      l'en-tete `Retry-After` (secondes).
    - Les listes paginees suivent le schema `Pagination` (parametres `page` et `per_page`).

    ## Authentification

    Le schema `bearerAuth` designe un jeton personnel Laravel Sanctum, envoye dans
    `Authorization: Bearer <id>|bdl_...`. Le prefixe `bdl_` (present apres le `|`) permet aux outils
    de detection de secrets de reperer un jeton divulgue. Les jetons expirent (7 jours par defaut).

    ## Anti-enumeration

    L'inscription, la demande de code de verification et la demande de reinitialisation de mot de
...
    erreur `422` (`invalid_verification_code`) pour un code inconnu, expire, deja utilise ou
...

...

...
servers:
...
paths:
...
  /me:
    delete:
      operationId: deleteAccount
      summary: Supprimer son compte
      description: >-
        Supprime definitivement le compte apres verification du mot de passe : les contacts, le
        profil, les demandes, les notifications et tous les jetons (y compris celui de la requete)
        sont effaces, et le compte est anonymise (statut `deleted`). Les correspondances en cours
        prennent fin ; les messages deja envoyes restent visibles par leur destinataire. Meme
        limite de debit qu'un changement de mot de passe.
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, format: password, maxLength: 255 }
      responses:
        '204':
          description: Compte supprime ; le jeton utilise n'est plus valide.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Laravel Sanctum personal access token (prefixe bdl_)
      description: >-
        Jeton personnel Sanctum envoye dans `Authorization: Bearer <jeton>`. Un jeton absent,
        invalide ou expire donne une reponse `401`.
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: >-
        Langue des messages. Valeurs prises en charge : `fr` et `ar`. Toute autre valeur est
        ignoree et la langue par defaut (`fr`) s'applique. La langue retenue est rappelee dans
        l'en-tete de reponse `Content-Language`.
      schema:
        type: string
        examples:
          - fr
          - ar
          - fr-FR,fr;q=0.9,ar;q=0.5
...
  headers:
    XRequestId:
      description: Identifiant de correlation genere par le serveur.
      schema:
        type: string
        format: uuid
    RetryAfter:
      description: Nombre de secondes a attendre avant de reessayer.
      schema:
        type: integer
        minimum: 0
    XRateLimitLimit:
      description: Limite de requetes pour la fenetre courante.
      schema:
        type: integer
    XRateLimitRemaining:
      description: Requetes restantes pour la fenetre courante.
      schema:
        type: integer
...
  responses:
...
    Unauthorized:
      description: Non authentifie (jeton absent, invalide ou expire).
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
        WWW-Authenticate:
          description: Schema d'authentification attendu.
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Unauthenticated.
            code: unauthenticated
            request_id: 0b5b7f1e-6f38-4e0c-9a0c-3f3c3f5e2a11
    Forbidden:
      description: Authentifie, mais action non autorisee.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: This action is unauthorized.
            code: forbidden
            request_id: 0b5b7f1e-6f38-4e0c-9a0c-3f3c3f5e2a11
...
    ValidationError:
      description: Donnees invalides. `errors` liste les messages par champ.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            message: The given data was invalid.
            code: validation_failed
            errors:
              name:
                - The name field is required.
            request_id: 0b5b7f1e-6f38-4e0c-9a0c-3f3c3f5e2a11
    TooManyRequests:
      description: Limite de debit depassee. Reessayer apres le delai indique par `Retry-After`.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-RateLimit-Limit:
          $ref: '#/components/headers/XRateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/XRateLimitRemaining'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Too many requests.
            code: too_many_requests
            request_id: 0b5b7f1e-6f38-4e0c-9a0c-3f3c3f5e2a11
    ServerError:
      description: >-
        Erreur interne. Le corps ne contient jamais de detail technique (pas de trace, de requete
        SQL ni de donnee personnelle) ; communiquer `request_id` au support.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Server error.
            code: server_error
            request_id: 0b5b7f1e-6f38-4e0c-9a0c-3f3c3f5e2a11
...
  schemas:
    ErrorCode:
      type: string
      description: >-
        Identifiant stable de l'erreur, a utiliser cote client a la place du message. Les codes
        `idempotency_*` ne concernent que les routes qui acceptent `Idempotency-Key`.
      enum:
        - bad_request
        - unauthenticated
        - forbidden
        - not_found
        - method_not_allowed
        - conflict
        - csrf_token_mismatch
        - validation_failed
        - too_many_requests
        - server_error
        - service_unavailable
        - http_error
        - idempotency_key_invalid
        - idempotency_key_reused
        - idempotency_request_in_progress
        - invalid_credentials
        - account_suspended
        - invalid_verification_code
        - account_not_suspended
        - invalid_request_transition
        - request_not_publishable
        - invalid_match_transition
        - contact_not_authorized
        - interaction_blocked
        - report_already_exists
        - invalid_report_transition
    Error:
      type: object
      required:
        - message
        - code
        - request_id
      properties:
        message:
          type: string
          description: Message lisible, traduit selon `Accept-Language`. Ne pas l'interpreter par programme.
        code:
          $ref: '#/components/schemas/ErrorCode'
        request_id:
          type: string
          format: uuid
          description: Identique a l'en-tete `X-Request-ID`.
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          required:
            - errors
          properties:
            code:
              const: validation_failed
            errors:
              type: object
              description: Messages de validation, par nom de champ.
              additionalProperties:
                type: array
                items:
                  type: string
...